+++ src/elements/content-sidebar/activity-feed/activity-feed/ActivityFeed.scss
@@ -115,6 +115,10 @@
         margin-top: 20px;
     }
 
+    .bcs-feed-item-inline-error {
+        margin: 20px 25px;
+    }
+
     .bcs-activity-feed-comment-input {
         flex: 0 0 auto;
         min-height: 93px;
